Fix Gradle subplugin deprecation
