+++ kotlin-power-assert-gradle/build.gradle.kts
@@ -11,9 +11,6 @@ plugins {
 dependencies {
   implementation(kotlin("stdlib"))
   implementation(kotlin("gradle-plugin-api"))
-
-  kapt("com.google.auto.service:auto-service:1.0-rc6")
-  compileOnly("com.google.auto.service:auto-service-annotations:1.0-rc6")
 }
 
 pluginBundle {
